+++ binx/registry.py
@@ -5,6 +5,7 @@ are created by the user and registered at runtime.
 
 from .exceptions import RegistryError
 from .utils import bfs_shortest_path
+import warnings
 
 _collection_registry = {}
 
@@ -16,16 +17,25 @@ def register_collection(cls):
     fullpath = cls.__module__ + '.' + cls.__name__
 
     if fullpath in _collection_registry: # if the fullpath is already
-        raise RegistryError('The classname {} has already been registered. Create a different name for your collection'.format(fullpath))
+        # if registered_adapters or adaptable_from already have values we should raise here
+        # this should be enough to ensure the adapter chain won't be compromised by accident
+        adapters = _collection_registry[fullpath][1]['registered_adapters']
+        from_colls = _collection_registry[fullpath][1]['adaptable_from']
 
-    # you must do the lookup by fullpath
-    _collection_registry[fullpath] = (cls, {
-        'serializer_class': cls.serializer_class,
-        'internal_class': cls.internal_class,
-        'registered_adapters': set(),  #NOTE these are the classes registered adapters
-        'adaptable_from': set()   #NOTE these are other collection objects a coll can be adapted from
-    })
-    #pprint(_collection_registry)
+        if len(adapters) > 0 or len(from_colls) > 0:
+            raise RegistryError('The name {} has already been registered in the adapter chain and cannot be overridden'.format(fullpath))
+
+        _collection_registry[fullpath][0] = cls
+        warnings.warn("The reference to {} has been overridden in the registry.".format(fullpath))
+
+    else:
+        # you must do the lookup by fullpath
+        _collection_registry[fullpath] = [cls, {
+            'serializer_class': cls.serializer_class,
+            'internal_class': cls.internal_class,
+            'registered_adapters': set(),  #NOTE these are the classes registered adapters
+            'adaptable_from': set()   #NOTE these are other collection objects a coll can be adapted from
+        }]
 
 
 def get_class_from_collection_registry(classname):